// src/api/generated/models/listAdminAppCasesRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ListAdminAppCasesRequest {
  /** @minimum 1 */
  pageNum?: number;
  /** @minimum 1 */
  pageSize?: number;
  status?: string;
  keyword?: string;
  featured?: boolean;
}