// src/api/generated/models/pageResultAdminAppCaseInfo.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AdminAppCaseInfo } from './adminAppCaseInfo';

export interface PageResultAdminAppCaseInfo {
  list?: AdminAppCaseInfo[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}